Challenge11 - Pheasant
Not very complicated problem. Of course you need to try keys until
proper decryption works out, but this can easily be checked since
the start of the decrypted data is already known. Another important
point, I think, is to build the feed for friends in the proper order,
merging results and stopping as soon as we have what we want (for
instance friends whose feed last time is before our last result
entry will not need to be used).