Challenge 15 - Take a corner
I simply started with this recursive backtracking solution and
found it was not so unefficient, so I just leave it this way. I
think it can be improved a lot but at least it's quite easy to
read and follow the algorithm. The Board::isValid method is
probably the ugliest because it implements all the logic of reversi
on its own, so readability can be improved starting here. Also
the Problem::solve method is a little large but I think its code
is readable enough :-)