Challenge 16 - ÑAPA
I have to say 'sorry' for my solution. About 3 seconds with the
test and around 3 hours on the submit phase (which I already
calculated). I tried to improve it during this 3h by distributing
particles in a R-Tree, and it was a quite promising solution (may
I someday finish it I could publish it here too), but submit ended
before and I just walked away.

In my defense I have to say that I conceived my official solution
in about 5 minutes or so (you know, that very silly idea of making
a 1000x1000 boxes grid drastically reduces complexity of the all-with-all
solution). So I think the best point of my solution is readability
and shortness.