Challenge 17 - The Failsystem
Well, an awful code for a beautiful problem, I guess.

I started modeling Sectors, Clusters, FATTable, DirEntries
and so, and also a bunch of code for printing them out to screen
both in hexadecimal and chars, as any binary data inspector, which
helped me to browse the important sectors.

After that I just replaced constants for those values I extracted
because they won't change, right? And deleted all that beautiful
stuff to add this big Disk class that is capable of getting the files
you ask for. So all the logic inside, a very ugly and improvable
solution from a maintainability point of view, ... but it worked.

Note that since the whole file weights ~5Gb some addresses can't be
accessed via a single fseek() operation. Even more, they can't be
represented as integers in php. Instead of using BC library to
use big numbers I just addressed the disk with a couple of integers
(sector no. and byte in sector) and properly performed cumulated
fseek() operations when needed.

Never gonna give you up...